Move mutex implementations to submodule
